+++ spec/regression/mimicks.spec.ts
@@ -1,6 +1,5 @@
 import test from 'ava'
-import { Substitute, Arg } from '../../src'
-import { mimicks } from '../../src/Transformations'
+import { Substitute, Arg, mimicks } from '../../src'
 
 interface Calculator {
   add(a: number, b: number): number
